relocated the testers for catheter box
